Move TurboActivityObserver to a new 'observers' package
